+++ packages/components/popover/src/popover.stories.ts
@@ -4,7 +4,6 @@ import type { Meta, StoryObj } from '@storybook/web-components';
 import '@sl-design-system/button/register.js';
 import '@sl-design-system/button-bar/register.js';
 import '@sl-design-system/text-field/register.js';
-import { anchor } from '@sl-design-system/shared';
 import { html } from 'lit';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import '../register.js';
@@ -21,24 +20,36 @@ export default {
 
 export const Basic: Story = {
   args: {
-    position: 'top'
+    position: 'bottom'
   },
   argTypes: {
     position: {
       control: 'inline-radio',
-      options: ['top', 'right', 'bottom', 'left']
+      options: [
+        'top',
+        'top-start',
+        'top-end',
+        'right',
+        'right-start',
+        'right-end',
+        'bottom',
+        'bottom-start',
+        'bottom-end',
+        'left',
+        'left-start',
+        'left-end'
+      ]
     }
   },
   render: ({ position }) => {
-    const onClick = (event: Event & { target: Button }): void => {
-      const popover = event.target.nextElementSibling as HTMLElement;
-
+    const onClick = (): void => {
+      const popover = document.querySelector('sl-popover') as HTMLElement;
       popover.togglePopover();
     };
 
     return html`
-      <sl-button @click=${onClick} id="button">Toggle popover</sl-button>
-      <sl-popover anchor="button" position=${ifDefined(position)}>I'm a popover</sl-popover>
+      <sl-button popovertarget="my-popover" @click=${onClick} id="button" variant="primary">Toggle popover</sl-button>
+      <sl-popover id="my-popover" anchor="button" position=${ifDefined(position)}> I'm a popover example </sl-popover>
     `;
   }
 };
@@ -46,25 +57,44 @@ export const Basic: Story = {
 export const All: Story = {
   render: () => {
     setTimeout(() => {
-      document.querySelectorAll('sl-popover').forEach(popover => popover.showPopover());
+      document.querySelectorAll('sl-popover').forEach(popover => {
+        popover.showPopover();
+      });
     });
 
     return html`
       <style>
-        sl-popover {
-          --sl-popover-offset: 16px;
+        div {
+          margin: 52px;
+          display: flex;
+          flex-direction: column;
+          align-items: center;
         }
       </style>
-      <span id="anchor">Popover</span>
-      <sl-popover anchor="anchor" popover="manual" position="top">Top</sl-popover>
-      <sl-popover anchor="anchor" popover="manual" position="right">Right</sl-popover>
-      <sl-popover anchor="anchor" popover="manual" position="bottom">Bottom</sl-popover>
-      <sl-popover anchor="anchor" popover="manual" position="left">Left</sl-popover>
+      <div>
+        <sl-button id="anchor" variant="primary">This is a popover anchor element (sl-button component) </br> with all top and bottom popover allowed positions shown <br> all examples at once</sl-button>
+        <sl-popover anchor="anchor" popover="manual" position="top">Top</sl-popover>
+        <sl-popover anchor="anchor" popover="manual" position="top-start">Top start</sl-popover>
+        <sl-popover anchor="anchor" popover="manual" position="top-end">Top end</sl-popover>
+        <sl-popover anchor="anchor" popover="manual" position="bottom">Bottom</sl-popover>
+        <sl-popover anchor="anchor" popover="manual" position="bottom-start">Bottom start</sl-popover>
+        <sl-popover anchor="anchor" popover="manual" position="bottom-end">Bottom end</sl-popover>
+      </div>
+
+      <div>
+        <sl-button id="anchor2" variant="primary" style="width: 72px">This is a popover anchor element (sl-button component) with all right and left popover allowed positions shown all examples at once</sl-button>
+        <sl-popover anchor="anchor2" popover="manual" position="right">Right <br> example</sl-popover>
+        <sl-popover anchor="anchor2" popover="manual" position="right-start">Right <br> start <br> example</sl-popover>
+        <sl-popover anchor="anchor2" popover="manual" position="right-end">Right <br> end <br> example</sl-popover>
+        <sl-popover anchor="anchor2" popover="manual" position="left">Left <br> example</sl-popover>
+        <sl-popover anchor="anchor2" popover="manual" position="left-start">Left <br> start <br> example</sl-popover>
+        <sl-popover anchor="anchor2" popover="manual" position="left-end">Left <br> end <br> example</sl-popover>
+      </div>
     `;
   }
 };
 
-export const Dialog: Story = {
+export const MoreComplexContent: Story = {
   render: () => {
     const onClick = (event: Event & { target: Button }): void => {
       const popover = event.target.nextElementSibling as HTMLElement;
@@ -72,56 +102,41 @@ export const Dialog: Story = {
       popover.togglePopover();
     };
 
+    const onClose = (event: Event & { target: Button }): void => {
+      event.target.closest('sl-popover')?.hidePopover();
+    };
+
     return html`
-      <!--added this to get the polyfill working in FF and Safari-->
       <style>
-        [popover],
-        :host(:where([popover])) {
-          background-color: canvas;
-          border-color: initial;
-          border-image: initial;
-          border-style: solid;
-          border-width: initial;
-          color: canvastext;
-          height: fit-content;
-          margin: auto;
-          overflow: auto;
-          padding: 0.25em;
-          position: fixed;
-          width: fit-content;
-          z-index: 2147483647;
+        header {
+          font: var(--sl-text-popover-text-title);
         }
 
-        [popover] {
-          inset: 0;
+        hr {
+          margin: 8px 0;
         }
 
-        [popover]:not(.\\:popover-open),
-        :host([popover]:not(.\\:popover-open)) {
-          display: none;
-        }
-
-        [popover]:is(dialog[open], .\\:popover-open) {
-          display: block;
-        }
-
-        [anchor].\\:popover-open {
-          inset: auto;
-        }
-
-        @supports selector([popover]:popover-open) {
-          [popover]:not(.\\:popover-open, dialog[open]),
-          :host([popover]:not(.\\:popover-open, dialog[open])) {
-            display: revert;
-          }
-
-          [anchor]:is(:popover-open) {
-            inset: auto;
-          }
+        footer {
+          display: flex;
+          gap: 8px;
+          justify-content: end;
         }
       </style>
-      <sl-button @click=${onClick} id="button">Toggle popover</sl-button>
-      <dialog anchor="button" popover ${anchor({ position: 'bottom' })}>I'm a popover</dialog>
+      <div>
+        <sl-button popovertarget="popover-1" id="anchor" variant="primary" @click=${onClick}>Toggle popover</sl-button>
+        <sl-popover id="popover-1" anchor="anchor">
+          <header>Please confirm</header>
+          <section>
+            <hr color="#D9D9D9" />
+            Are you sure you want to continue?
+            <hr color="#D9D9D9" />
+          </section>
+          <footer>
+            <sl-button @click=${onClose} autofocus size="sm">Cancel</sl-button>
+            <sl-button @click=${onClose} size="sm" variant="primary">Confirm</sl-button>
+          </footer>
+        </sl-popover>
+      </div>
     `;
   }
 };
@@ -163,41 +178,6 @@ export const Edges: Story = {
       <sl-popover anchor="anchor2" popover="manual" position="right">Right</sl-popover>
       <sl-popover anchor="anchor3" popover="manual" position="bottom">Bottom</sl-popover>
       <sl-popover anchor="anchor4" popover="manual" position="left">Left</sl-popover>
-    `;
-  }
-};
-
-export const Focus: Story = {
-  render: () => {
-    const onClick = (event: Event & { target: Button }): void => {
-      const popover = event.target.nextElementSibling as HTMLElement;
-
-      popover.togglePopover();
-    };
-
-    return html`
-      <style>
-        sl-popover form {
-          align-items: center;
-          display: grid;
-          gap: 0.5rem;
-          grid-template-columns: auto 1fr;
-          grid-template-rows: 1fr 1fr;
-        }
-        sl-button-bar {
-          grid-column: 1 / -1;
-        }
-      </style>
-      <sl-button @click=${onClick} id="button">Toggle popover</sl-button>
-      <sl-popover anchor="button">
-        <form>
-          <label>Label</label>
-          <sl-text-field placeholder="Input"></sl-text-field>
-          <sl-button-bar align="end">
-            <sl-button size="sm">Save</sl-button>
-          </sl-button-bar>
-        </form>
-      </sl-popover>
     `;
   }
 };